Port upstream pdfbox-jbig2 correctness fixes (arithmetic decoding, streams, refinement region)
Ports several bug fixes from apache/pdfbox-jbig2 that landed after this port
was originally taken (PDFBOX-6151 through PDFBOX-6199):

- ArithmeticIntegerDecoder: PREV was a shared mutable field, so interleaved
  calls to Decode()/DecodeIAID() (e.g. via IAID decoding during text region
  refinement) could corrupt each other's state; made PREV a local variable
  threaded through the calls, matching upstream. Also masks the IAID context
  index to the correct SBSYMCODELEN bits instead of letting it grow past the
  allocated CX size.
- CX: added Copy() for deep-copying context state, needed when contexts are
  shared/reused across segments.
- AbstractImageInputStream/SubInputStream: stream position could be advanced
  past EOF (or decremented, since a Java-style read()==-1 leaked into a
  position increment), and SkipBits() could push the position beyond the
  stream length.
- PatternDictionary: HdpWidth/HdpHeight were read as sign-extended sbyte
  instead of an unsigned 0-255 value, corrupting collective bitmap width for
  gray-code values >= 128.
- GenericRefinementRegion: Template 1's SLTP context constant was 0x080
  instead of the correct 0x008 (Figure 15), and the AT-pixel override check
  had inverted logic (enabled override only when the AT pixel matched
  non-nominal placement, instead of catching the non-nominal case). Also
  replaced the byte-packed Template 1 decode paths (never validated against
  the spec and superseded upstream) with a straightforward pixel-by-pixel
  implementation, and added support for refining directly against the page
  buffer when a refinement segment has no referred-to segment (7.4.7.4).

+++ UglyToad.PdfPig.Filters.Jbig2.PdfboxJbig2/Jbig2/PatternDictionary.cs
@@ -72,8 +72,8 @@
 
         private void ReadPatternWidthAndHeight()
         {
-            HdpWidth = (sbyte)subInputStream.ReadByte();
-            HdpHeight = (sbyte)subInputStream.ReadByte();
+            HdpWidth = (short)subInputStream.ReadByte();
+            HdpHeight = (short)subInputStream.ReadByte();
         }
 
         private void ReadGrayMax()
